requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
kombu==5.4.2
multidict==6.1.0
numpy==2.2.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pillow==11.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
python-telegram-bot==21.6
redis==5.2.0
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.40.0
tenacity==9.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.15.4
